requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
streamlit>=1.31.0
plotly>=5.18.0
tenacity>=8.2.0
tqdm>=4.66.0